Site form

+++ packages/model/repositories/business-site.ts
@@ -2,6 +2,13 @@ import { BaseRepository } from "../lib/base-repository";
 import { deleteFile, uploadFile } from "../lib/upload_file";
 import { Prisma } from "../prisma/prisma-client";
 import { BusinessSiteModel, CompleteBusinessSite } from "../prisma/zod";
+import {
+  BusinessSiteUpdateValidation,
+  BusinessSiteValidation,
+} from "../validation/business-site";
+
+const businessImageFolder = (businessId: string) =>
+  `business_sites/${businessId}`;
 
 export class BusinessSiteRepository extends BaseRepository<
   CompleteBusinessSite,
@@ -11,10 +18,15 @@ export class BusinessSiteRepository extends BaseRepository<
     super(BusinessSiteModel.omit({ id: true }), "businessSite");
   }
 
+  protected init() {
+    this.addValidator("create", BusinessSiteValidation);
+    this.addValidator("update", BusinessSiteUpdateValidation);
+  }
+
   protected doCreate(data: any) {
     const { businessId, ...rest } = data;
     return uploadFile(
-      `business_sites/${businessId}`,
+      businessImageFolder(businessId),
       rest.logo,
       async (blob) => {
         const entity = await super.doCreate({
@@ -28,21 +40,26 @@ export class BusinessSiteRepository extends BaseRepository<
   }
 
   protected async doUpdate(id: string, data: any) {
-    const { businessId, ...rest } = data;
+    const { businessId, logo, ...rest } = data;
     const originalEntity = await super.getById(id);
-    return uploadFile(
-      `business_sites/${businessId}`,
-      rest.logo,
-      async (blob) => {
-        const entity = await super.doUpdate(id, {
-          ...this.getObject(rest),
-          businessId,
-          logo: blob,
-        });
+    return uploadFile(businessImageFolder(businessId), logo, async (blob) => {
+      const newData = {
+        ...this.getObject(rest),
+        businessId,
+      };
+      if (blob) {
+        newData.logo = blob;
+      }
+      const entity = await super.doUpdate(id, newData);
+      if (logo && originalEntity.logo) {
         await deleteFile(originalEntity.logo.url);
-        return entity;
-      },
-    );
+      }
+      return entity;
+    });
+  }
+
+  getByBusinessId(businessId: string) {
+    return this.model.findUnique({ where: { businessId } });
   }
 }
